Create useSpotify hook. Create use-spotify component
Refactored the spotify user data components in to one. It currently shows the data on the homepage once the user logs in, but we will change this to display on the user profile after login
