+++ src/types/index.ts
@@ -1,0 +1,12 @@
+export interface Artist {
+  name: string
+  image: string
+  spotifyUrl: string
+}
+
+export interface UserProfile {
+  name: string
+  email: string
+  id: string
+  userImage: string | null
+}
